Allow launching Karma tests with Firefox (#6)
Many candidates (especially those using Linux, but on other operating
systems too) only have Firefox installed. But the bug reproduces in
Firefox as well as Chrome, so it doesn't make sense to force them to
install Chrome if they'd prefer not to.

This checks the relevant files into `node_modules` and adds a
commented-out Karma config for using Firefox.

I tested that this works on my personal Linux PC which does not have
Chrome installed, and verified that Karma was able to run in both
single-run and watch mode.